colorize: fix compilation with -Wall
Fix the following error when running `make validate`:

 filters/colorize.c:82:19: error: conversion to ‘size_t’ {aka ‘long
 unsigned int’} from ‘int’ may change the sign of the result
 [-Werror=sign-conversion]
    82 |                 n += res; \
       |                   ^~

That required cast was lost during the applying the original patch.

Fixes: 899f1699f841 ("colorize: use snprintf instead of sprintf")
Acked-by: Robin Jarry <robin@jarry.cc>

+++ filters/colorize.c
@@ -79,7 +79,7 @@ static const char *seq(struct style *s) {
 			perror("fatal: failed to format sequence"); \
 			abort(); \
 		} \
-		n += res; \
+		n += (size_t)res; \
 	} while (0)
 
 		XSPRINTF("%s", "\x1b[");
